refactor code for prettier reading
